pyproject.toml
[project]
name = "trabalho-bd-backend"
version = "0.1.0"
description = "Backend FastAPI para o projeto AirbnbRJ - Trabalho de Banco de Dados"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "mysql-connector-python>=9.1.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []
